+++ archive.php
@@ -9,19 +9,22 @@
 
 get_header(); ?>
 
-	<div id="primary" class="content-area">
-		<main id="main" class="site-main">
-
 		<?php
 		if ( have_posts() ) : ?>
 
-			<header class="page-header">
-				<?php
-					the_archive_title( '<h1 class="page-title">', '</h1>' );
-					the_archive_description( '<div class="archive-description">', '</div>' );
-				?>
-			</header><!-- .page-header -->
+		<section class="page-top section-space">
 
+		  <div class="container">
+		    <div class="row">
+		      <div class="col-md-12 section-title">
+				<?php
+					the_archive_title( '<h3 class="page-title">', '</h3>' );
+					the_archive_description( '<small class="archive-description">', '</small>' );
+				?>
+				</div>
+			</div>
+
+<main class="row" id="primary">
 			<?php
 			/* Start the Loop */
 			while ( have_posts() ) : the_post();
@@ -42,10 +45,9 @@ get_header(); ?>
 			get_template_part( 'template-parts/content', 'none' );
 
 		endif; ?>
+</main>
+</div><!--.container-->
+</section>
 
-		</main><!-- #main -->
-	</div><!-- #primary -->
 
-<?php
-get_sidebar();
-get_footer();
+<?php get_footer(); ?>